Wrapper script for wget/curl
Script is capable of downloading file based on specified URL and also
use optionally specified file name as a name of the output.

Most important feature is that it creates a log file that contains
information specified on the command line and also generated checksums.
In addition to that it knows how to replay it.