Bootstraping code for executing generated scripts
Most of the time people want shell scripts to be simple and transparent,
but there is a domain where this is not one of the main requriements. In
such cases one might consider creating script that generates another
script, or other kind of executable, and then executes it. One of such
use cases is when the result has to be optimized for specific platform
or Linux distribution.

This commit contains simple function that does allow just that and with
minimal requirements from the system it is running on. Only thing left
for the developer of such script is to define generateScript function
that will create the content of executable that should be executed. Hello
World example:

    #!/bin/bash

    set -e

    function generateScript()
    {
        cat << EOF
    trap "rm -f -- '\$0'" INT TERM EXIT
    echo 'Hello World!'
    EOF
    }

    source "<install-prefix>/snippets/scripts/functions/bootstrap.sh"

Alternative way to define the above code is to inline the content of
"bootstrap.sh" which should be prefered in cases that the result is
deployed somewhere.