Setup batch file for windows
Batch file initializes environment so that batch files from this
repository can be easily used on computer with Windows OS.

It creates directory C:\opt if it doesn't exist and in it it creates
symbolic link named batch-files that is pointing to its working directory.
As a last step it merges registry entires from registry-entries.reg that
also is looked for in its working directory.