Look for bashrc components in repository
If ~/.bashrc is a symbolic link, then it tries to find bashrc components
in the same directory where it points to. This simplifies installation
to just creating symbolic link to dot.bashrc and it will find the rest
by its self. Additional feature is that now it's possible to define
~/.bash_${sfx} components separately and those will be sourced as well.
It's very useful on systems where it is necessary to have specific
definitions.